remove need to restart when (dis)enabling
Along the way restructure the package so that
codemirror keybindings are kept separated from
jlab command mode bindings.
