+++ src/labCommands.ts
@@ -1,0 +1,457 @@
+import { JupyterFrontEnd } from '@jupyterlab/application';
+import { MarkdownCell } from '@jupyterlab/cells';
+import { CodeMirrorEditor } from '@jupyterlab/codemirror';
+import {
+  INotebookTracker,
+  NotebookActions,
+  NotebookPanel
+} from '@jupyterlab/notebook';
+import { ReadonlyPartialJSONObject } from '@lumino/coreutils';
+
+import { IDisposable } from '@lumino/disposable';
+import { ElementExt } from '@lumino/domutils';
+
+export function addJLabCommands(
+  app: JupyterFrontEnd,
+  tracker: INotebookTracker,
+  CodeMirror: CodeMirrorEditor
+): Array<IDisposable> {
+  const { commands, shell } = app;
+  function getCurrent(args: ReadonlyPartialJSONObject): NotebookPanel | null {
+    const widget = tracker.currentWidget;
+    const activate = args['activate'] !== false;
+
+    if (activate && widget) {
+      shell.activateById(widget.id);
+    }
+
+    return widget;
+  }
+  function isEnabled(): boolean {
+    return (
+      tracker.currentWidget !== null &&
+      tracker.currentWidget === app.shell.currentWidget
+    );
+  }
+  const addedCommands = [
+    commands.addCommand('run-select-next-edit', {
+      label: 'Run Cell and Edit Next Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { context, content } = current;
+          NotebookActions.runAndAdvance(content, context.sessionContext);
+          current.content.mode = 'edit';
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('run-cell-and-edit', {
+      label: 'Run Cell and Edit Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { context, content } = current;
+          NotebookActions.run(content, context.sessionContext);
+          current.content.mode = 'edit';
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('cut-cell-and-edit', {
+      label: 'Cut Cell(s) and Edit Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          NotebookActions.cut(content);
+          content.mode = 'edit';
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('copy-cell-and-edit', {
+      label: 'Copy Cell(s) and Edit Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          NotebookActions.copy(content);
+          content.mode = 'edit';
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('paste-cell-and-edit', {
+      label: 'Paste Cell(s) and Edit Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          NotebookActions.paste(content, 'below');
+          content.mode = 'edit';
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('merge-and-edit', {
+      label: 'Merge and Edit Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          NotebookActions.mergeCells(content);
+          current.content.mode = 'edit';
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('enter-insert-mode', {
+      label: 'Enter Insert Mode',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          if (content.activeCell !== null) {
+            const editor = content.activeCell.editor as CodeMirrorEditor;
+            current.content.mode = 'edit';
+            (CodeMirror as any).Vim.handleKey(editor.editor, 'i');
+          }
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('leave-insert-mode', {
+      label: 'Leave Insert Mode',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          if (content.activeCell !== null) {
+            const editor = content.activeCell.editor as CodeMirrorEditor;
+            (CodeMirror as any).Vim.handleKey(editor.editor, '<Esc>');
+          }
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('select-below-execute-markdown', {
+      label: 'Execute Markdown and Select Cell Below',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          if (
+            content.activeCell !== null &&
+            content.activeCell.model.type === 'markdown'
+          ) {
+            (current.content.activeCell as MarkdownCell).rendered = true;
+          }
+          return NotebookActions.selectBelow(current.content);
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('select-above-execute-markdown', {
+      label: 'Execute Markdown and Select Cell Below',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          if (
+            content.activeCell !== null &&
+            content.activeCell.model.type === 'markdown'
+          ) {
+            (current.content.activeCell as MarkdownCell).rendered = true;
+          }
+          return NotebookActions.selectAbove(current.content);
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('select-first-cell', {
+      label: 'Select First Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          content.activeCellIndex = 0;
+          content.deselectAll();
+          if (content.activeCell !== null) {
+            ElementExt.scrollIntoViewIfNeeded(
+              content.node,
+              content.activeCell.node
+            );
+          }
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('select-last-cell', {
+      label: 'Select Last Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current) {
+          const { content } = current;
+          content.activeCellIndex = current.content.widgets.length - 1;
+          content.deselectAll();
+          if (content.activeCell !== null) {
+            ElementExt.scrollIntoViewIfNeeded(
+              content.node,
+              content.activeCell.node
+            );
+          }
+        }
+      },
+      isEnabled
+    }),
+    commands.addCommand('center-cell', {
+      label: 'Center Cell',
+      execute: args => {
+        const current = getCurrent(args);
+
+        if (current && current.content.activeCell !== null) {
+          const er = current.content.activeCell.inputArea.node.getBoundingClientRect();
+          current.content.scrollToPosition(er.bottom, 0);
+        }
+      },
+      isEnabled
+    }),
+
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'U'],
+      command: 'notebook:undo-cell-action'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', '-'],
+      command: 'notebook:split-cell-at-cursor'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'D'],
+      command: 'cut-cell-and-edit'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'Y'],
+      command: 'copy-cell-and-edit'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'P'],
+      command: 'paste-cell-and-edit'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl Shift J'],
+      command: 'notebook:extend-marked-cells-below'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Ctrl Shift J'],
+      command: 'notebook:extend-marked-cells-below'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl Shift K'],
+      command: 'notebook:extend-marked-cells-above'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Ctrl Shift K'],
+      command: 'notebook:extend-marked-cells-above'
+    }),
+    // this one doesn't work yet
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'Shift O'],
+      command: 'notebook:insert-cell-above'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'Ctrl O'],
+      command: 'notebook:insert-cell-above'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'O'],
+      command: 'notebook:insert-cell-below'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl J'],
+      command: 'select-below-execute-markdown'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl K'],
+      command: 'select-above-execute-markdown'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Escape'],
+      command: 'leave-insert-mode'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl ['],
+      command: 'leave-insert-mode'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Ctrl I'],
+      command: 'enter-insert-mode'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl Enter'],
+      command: 'run-cell-and-edit'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Shift Enter'],
+      command: 'run-select-next-edit'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Shift Escape'],
+      command: 'notebook:enter-command-mode'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Shift M'],
+      command: 'merge-and-edit'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Accel 1'],
+      command: 'notebook:change-cell-to-code'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Accel 2'],
+      command: 'notebook:change-cell-to-markdown'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Accel 3'],
+      command: 'notebook:change-cell-to-raw'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'G'],
+      command: 'select-first-cell'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'Ctrl G'],
+      command: 'select-last-cell'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['G', 'G'],
+      command: 'select-first-cell'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Shift G'],
+      command: 'select-last-cell'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Y', 'Y'],
+      command: 'notebook:copy-cell'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['D', 'D'],
+      command: 'notebook:cut-cell'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Shift P'],
+      command: 'notebook:paste-cell-above'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['P'],
+      command: 'notebook:paste-cell-below'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['O'],
+      command: 'notebook:insert-cell-below'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Shift O'],
+      command: 'notebook:insert-cell-above'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['U'],
+      command: 'notebook:undo-cell-action'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Ctrl E'],
+      command: 'notebook:move-cell-down'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Ctrl Y'],
+      command: 'notebook:move-cell-up'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Z', 'Z'],
+      command: 'center-cell'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Z', 'C'],
+      command: 'notebook:hide-cell-code'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Z', 'O'],
+      command: 'notebook:show-cell-code'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Z', 'M'],
+      command: 'notebook:hide-all-cell-code'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook:focus',
+      keys: ['Z', 'R'],
+      command: 'notebook:show-all-cell-code'
+    }),
+    commands.addKeyBinding({
+      selector: '.jp-Notebook.jp-mod-editMode',
+      keys: ['Ctrl O', 'Z', 'Z'],
+      command: 'center-cell'
+    }),
+    commands.addKeyBinding({
+      selector:
+        '.jp-Notebook.jp-mod-editMode .jp-InputArea-editor:not(.jp-mod-has-primary-selection)',
+      keys: ['Ctrl G'],
+      command: 'tooltip:launch-notebook'
+    })
+  ];
+  return addedCommands;
+}
